Restructure project layout
Follow standards from github.com/golang-standards/project-layout
