+++ internal/app/fipcontroller/controller.go
@@ -2,56 +2,31 @@ package fipcontroller
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"github.com/sirupsen/logrus"
-	"io/ioutil"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/hetznercloud/hcloud-go/hcloud"
-
+	"github.com/sirupsen/logrus"
 	"k8s.io/client-go/kubernetes"
+
+	"github.com/cbeneke/hcloud-fip-controller/internal/pkg/configuration"
 )
-
-type stringArrayFlags []string
-
-func (flags *stringArrayFlags) String() string {
-	return fmt.Sprintf("['%s']", strings.Join(*flags, "', '"))
-}
-
-func (flags *stringArrayFlags) Set(value string) error {
-	*flags = append(*flags, value)
-	return nil
-}
-
-type Configuration struct {
-	HcloudApiToken    string           `json:"hcloud_api_token,omitempty"`
-	HcloudFloatingIPs stringArrayFlags `json:"hcloud_floating_ips,omitempty"`
-	LeaseDuration     int              `json:"lease_duration,omitempty"`
-	LeaseName         string           `json:"lease_name,omitempty"`
-	Namespace         string           `json:"namespace,omitempty"`
-	NodeAddressType   string           `json:"node_address_type,omitempty"`
-	NodeName          string           `json:"node_name,omitempty"`
-	PodName           string           `json:"pod_name,omitempty"`
-	LogLevel          string           `json:"log_level,omitempty"`
-}
 
 type Controller struct {
 	HetznerClient    *hcloud.Client
 	KubernetesClient *kubernetes.Clientset
-	Configuration    *Configuration
+	Configuration    *configuration.Configuration
 	Logger           *logrus.Logger
 }
 
-func NewController(config *Configuration) (*Controller, error) {
-	hetznerClient, err := hetznerClient(config.HcloudApiToken)
+func NewController(config *configuration.Configuration) (*Controller, error) {
+	hetznerClient, err := newHetznerClient(config.HcloudApiToken)
 	if err != nil {
 		return nil, fmt.Errorf("could not initialise hetzner client: %v", err)
 	}
 
-	kubernetesClient, err := kubernetesClient()
+	kubernetesClient, err := newKubernetesClient()
 	if err != nil {
 		return nil, fmt.Errorf("could not initialise kubernetes client: %v", err)
 	}
@@ -78,49 +53,6 @@ func NewController(config *Configuration) (*Controller, error) {
 	}, nil
 }
 
-// Read given config file and overwrite options from given Configuration
-func (configuration *Configuration) VarsFromFile(configFile string) error {
-	file, err := ioutil.ReadFile(configFile)
-	if err != nil {
-		return fmt.Errorf("failed to read config file: %v", err)
-	}
-
-	err = json.Unmarshal(file, &configuration)
-	if err != nil {
-		return fmt.Errorf("failed to decode config file: %v", err)
-	}
-
-	return nil
-}
-
-func (configuration *Configuration) Validate() error {
-	var errs []string
-	var undefinedErrs []string
-
-	if configuration.HcloudApiToken == "" {
-		undefinedErrs = append(errs, "hetzner cloud API token")
-	}
-	if len(configuration.HcloudFloatingIPs) <= 0 {
-		undefinedErrs = append(errs, "hetzner cloud floating IPs")
-	}
-	if configuration.NodeName == "" {
-		undefinedErrs = append(errs, "kubernetes node name")
-	}
-	if configuration.Namespace == "" {
-		undefinedErrs = append(errs, "kubernetes namespace")
-	}
-	if configuration.LeaseDuration <= 0 {
-		errs = append(errs, "lease duration needs to be greater than one")
-	}
-
-	if len(undefinedErrs) > 0 {
-		errs = append(errs, fmt.Sprintf("required configuration options not configured: %s", strings.Join(errs, ", ")))
-	}
-	if len(errs) > 0 {
-		return fmt.Errorf("%s", strings.Join(errs, ", "))
-	}
-	return nil
-}
 
 func (controller *Controller) Run(ctx context.Context) error {
 	if err := controller.UpdateFloatingIPs(ctx); err != nil {
