+++ internal/app/fipcontroller/leaderelection.go
@@ -2,63 +2,13 @@ package fipcontroller
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"strings"
 	"time"
 
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/leaderelection"
 	"k8s.io/client-go/tools/leaderelection/resourcelock"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func kubernetesClient() (*kubernetes.Clientset, error) {
-	kubeConfig, err := rest.InClusterConfig()
-	if err != nil {
-		return nil, fmt.Errorf("could not get kubeconfig: %v", err)
-	}
-
-	kubernetesClient, err := kubernetes.NewForConfig(kubeConfig)
-	if err != nil {
-		return nil, fmt.Errorf("could not get kubernetes client: %v", err)
-	}
-
-	return kubernetesClient, nil
-}
-
-func (controller *Controller) nodeAddress(nodeName, nodeAddressType string) (address net.IP, err error) {
-	nodes, err := controller.KubernetesClient.CoreV1().Nodes().List(metav1.ListOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("could not list nodes: %v", err)
-	}
-	controller.Logger.Debugf("Found %d nodes", len(nodes.Items))
-
-	var addresses []corev1.NodeAddress
-	for _, node := range nodes.Items {
-		if node.Name == nodeName {
-			addresses = node.Status.Addresses
-			break
-		}
-	}
-	controller.Logger.Debugf("Found %d addresses", len(addresses))
-
-	checkAddressType := corev1.NodeExternalIP
-	if nodeAddressType == "internal" {
-		checkAddressType = corev1.NodeInternalIP
-	}
-	controller.Logger.Debugf("Using address type %s", checkAddressType)
-
-	for _, address := range addresses {
-		if address.Type == checkAddressType {
-			return net.ParseIP(address.Address), nil
-		}
-	}
-	return nil, fmt.Errorf("could not find address for node %s", nodeName)
-}
 
 func (controller *Controller) leaseLock(id string) (lock *resourcelock.LeaseLock) {
 	lock = &resourcelock.LeaseLock{
@@ -110,3 +60,4 @@ func (controller *Controller) onStartedLeading(ctx context.Context) {
 func (controller *Controller) onStoppedLeading() {
 	controller.Logger.Info("Stopped leading")
 }
+
